requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
pillow>=10.0.0
scikit-learn>=1.3.0
prophet>=1.1.5
tensorflow>=2.13.0
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
